Distribute publishing of values across loop invocations
Publishing all sensor value updates in one loop invocations, even
with yields and the code taking less than 5ms to execute, causes
the api connection to Home Assistant to be very unstable on
ESP8266. With this hack it is much more stable, but still the
connection is dropped once in a while.